update validation in menu and delete items

+++ main.py
@@ -23,6 +23,7 @@ import cashier
 
 
 def menu():
+    error = False
     # Menampilkan menu utama pada console
     while True:
         # Membersihkan console dan menampilkan menu
@@ -30,29 +31,40 @@ def menu():
         print(f"========== ========== ========== ========== ==========\n")
         print(f"                    Super Cashier")
         print(f"\n========== ========== ========== ========== ==========\n")
+        if error == True:
+            print("Pilihan tidak ada pada menu")
+            print(f"\n========== ========== ========== ========== ==========\n")
+            error = False
         print(f"1. Buat Transaksi")
         print(f"0. Keluar")
         print(f"\n========== ========== ========== ========== ==========\n")
 
-        # Input pilihan user
-        pilihan = int(input("Masukkan Pilihan: "))
+        try:
+            # Input pilihan user
+            pilihan = int(input("Masukkan Pilihan: "))
 
-        # Case sesuai pilihan user
-        match pilihan:
-            case 1:
-                trnsct_123 = cashier.Transaction()
-                submenu(trnsct_123)
-            case 0:
-                print("See You")
-                print(f"\n========== ========== ========== ========== ==========\n")
-                exit()
-            case default:
-                print("Pilihan tidak ada pada menu")
-                print(f"\n========== ========== ========== ========== ==========\n")
+            if type(pilihan) != int:
+                raise ValueError
+            else:
+                # Case sesuai pilihan user
+                match pilihan:
+                    case 1:
+                        trnsct_123 = cashier.Transaction()
+                        submenu(trnsct_123)
+                    case 0:
+                        print("See You")
+                        print(
+                            f"\n========== ========== ========== ========== ==========\n")
+                        exit()
+                    case _:
+                        raise ValueError
+        except ValueError:
+            error = True
 
 
 def submenu(trnsct_123):
     # Menampilkan sub-menu pada console
+    error = False
     case_7 = False
     while True:
         # Membersihkan console dan menampilkan menu
@@ -60,6 +72,10 @@ def submenu(trnsct_123):
         print(f"========== ========== ========== ========== ==========\n")
         print(f"                    Super Cashier")
         print(f"\n========== ========== ========== ========== ==========\n")
+        if error == True:
+            print("Pilihan tidak ada pada menu")
+            print(f"\n========== ========== ========== ========== ==========\n")
+            error = False
         trnsct_123.check_order()
         if case_7 == True:
             trnsct_123.total_price()
@@ -74,29 +90,35 @@ def submenu(trnsct_123):
         print(f"0. Kembali")
         print(f"\n========== ========== ========== ========== ==========\n")
 
-        # Input pilihan user
-        pilihan = int(input("Masukkan Pilihan: "))
+        try:
+            # Input pilihan user
+            pilihan = int(input("Masukkan Pilihan: "))
 
-        # Case sesuai pilihan user
-        match pilihan:
-            case 1:
-                trnsct_123.add_item()
-            case 2:
-                trnsct_123.update_item_name()
-            case 3:
-                trnsct_123.update_item_qty()
-            case 4:
-                trnsct_123.update_item_price()
-            case 5:
-                trnsct_123.delete_item()
-            case 6:
-                trnsct_123.reset_transaction()
-            case 7:
-                case_7 = True
-            case 0:
-                menu()
-            case default:
-                print("Pilihan tidak ada pada menu")
+            if type(pilihan) != int:
+                raise ValueError
+            else:
+                # Case sesuai pilihan user
+                match pilihan:
+                    case 1:
+                        trnsct_123.add_item()
+                    case 2:
+                        trnsct_123.update_item_name()
+                    case 3:
+                        trnsct_123.update_item_qty()
+                    case 4:
+                        trnsct_123.update_item_price()
+                    case 5:
+                        trnsct_123.delete_item()
+                    case 6:
+                        trnsct_123.reset_transaction()
+                    case 7:
+                        case_7 = True
+                    case 0:
+                        menu()
+                    case _:
+                        raise ValueError
+        except ValueError:
+            error = True
 
 
 if __name__ == "__main__":
